Crate re_redap_browser

This crates implements the Redap browser feature, including the communication and UI aspects of it.

Modules

add_server_modal 🔒

context 🔒

entries 🔒

servers 🔒

tables_session_context 🔒

Helper to maintain a [SessionContext] with the tables of a remote server.

Structs

RedapServers

All servers known to the viewer, and their catalog data.

SortDatasetsResults

Enums

EntryKind

Statics

EXAMPLES_ORIGIN

Origin used to show the examples ui in the redap browser.

LOCAL_ORIGIN

Origin used to show the local ui in the redap browser.

Functions

dataset_and_its_recordings_ui

sort_datasets